% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_endpoint_url}
\alias{build_endpoint_url}
\title{Build URL to an ohsome API endpoint}
\usage{
build_endpoint_url(endpoint)
}
\arguments{
\item{endpoint}{character (atomic or vector)}
}
\value{
character
}
\description{
Build URL to an ohsome API endpoint
}
\keyword{Internal}
