% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_content}
\alias{convert_content}
\title{Convert ohsome metadata content}
\usage{
convert_content(content)
}
\arguments{
\item{content}{The parsed content of a response from the metadata endpoint of
ohsome API}
}
\value{
A list (parsed and converted content of ohsome metadata)
}
\description{
Converts the following elements: \code{apiVersion} to \code{numeric_version},
\code{spatialExtent} to \code{sfc_POLYGON} and \code{temporalExtent} to a
vector of \code{POSIXct}
}
\keyword{Internal}
