% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_quietly}
\alias{convert_quietly}
\title{Type convert without message}
\usage{
convert_quietly(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Converts type of data.frame columns with \code{readr::type_convert()} while
suppressing messages
}
\keyword{Internal}
