% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_spatialExtent}
\alias{convert_spatialExtent}
\title{Convert spatialExtent of ohsome metadata}
\usage{
convert_spatialExtent(spatialExtent)
}
\arguments{
\item{spatialExtent}{The \code{$extractRegion$spatialExtent} element of the
parsed content of a response from the metadata endpoint of ohsome API}
}
\value{
An \code{sfc_POLYGON} object
}
\description{
Convert spatialExtent of ohsome metadata
}
\keyword{internal}
