% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_temporalExtent}
\alias{convert_temporalExtent}
\title{Convert temporalExtent of ohsome metadata}
\usage{
convert_temporalExtent(temporalExtent)
}
\arguments{
\item{temporalExtent}{The \code{$extractRegion$temporalExtent} element of the
parsed content of a response from the metadata endpoint of ohsome API}
}
\value{
A named POSIXct vector
}
\description{
Convert temporalExtent of ohsome metadata
}
\keyword{internal}
