% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_metadata_message}
\alias{create_metadata_message}
\title{Create metadata message}
\usage{
create_metadata_message(meta)
}
\arguments{
\item{meta}{An ohsome_metadata object}
}
\value{
character
}
\description{
Creates a message text from a \code{ohsome_metadata} object.
}
\keyword{Internal}
