% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_endpoint}
\alias{extract_endpoint}
\title{Extract endpoint}
\usage{
extract_endpoint(query)
}
\arguments{
\item{query}{An \code{ohsome_query} object constructed with \code{\link[=ohsome_query]{ohsome_query()}} or
any of its wrapper functions}
}
\value{
character
}
\description{
Extracts the API endpoint path from the URL in an \code{ohsome_query} object
}
\keyword{Internal}
