% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null coalesce operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{left-side argument}

\item{b}{right-side argument}
}
\description{
Operator that returns the left-hand side operand if it is not \code{NULL}, else
returns the right-hand side operand
}
\keyword{Internal}
