% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_aggregate_elements.R
\name{ohsome_aggregate_elements}
\alias{ohsome_aggregate_elements}
\title{Aggregate elements}
\usage{
ohsome_aggregate_elements(
  boundary,
  aggregation = c("count", "length", "perimeter", "area"),
  groupBy = c("none", "boundary", "key", "tag", "type"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for OSM element aggregation
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#elements-aggregation}
}
