% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_contributions_count.R
\name{ohsome_contributions_count}
\alias{ohsome_contributions_count}
\title{Count OSM contributions}
\usage{
ohsome_contributions_count(
  boundary = NULL,
  latest = FALSE,
  return_value = c("absolute", "density"),
  time = lubridate::format_ISO8601(.ohsome_temporalExtent),
  ...
)
}
\arguments{
\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{latest}{logical; if true, request the count of only the latest
contributions provided to the OSM data.}

\item{return_value}{character; the value to be returned by the ohsome API:
\describe{
\item{absolute}{Returns the absolute number of contributions (default).}
\item{density}{Returns the number of contributions per square kilometer.}
}}

\item{time}{character; time parameter of the query (see
\href{https://docs.ohsome.org/ohsome-api/v1/time.html}{Supported time formats};
defaults to the temporal extent of the underlying OSHDB)}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for OSM contributions count
}
\details{
\code{ohsome_contributions_count} creates an \code{ohsome_query} object for
OSM element aggregation. Boundary objects are passed via \code{\link{set_boundary}} into
\code{\link{ohsome_boundary}}.
}
\examples{

ohsome_contributions_count("0,0,10", filter = "man_made=*", time = "2010/2020/P1Y")
ohsome_contributions_count("0,0,10", latest = TRUE, filter = "man_made=*", time = "2010/2020/P1Y")

}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#users-aggregation}
}
