% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_extract_contributions.R
\name{ohsome_extract_contributions}
\alias{ohsome_extract_contributions}
\alias{ohsome_contributions_bbox}
\alias{ohsome_contributions_centroid}
\alias{ohsome_contributions_geometry}
\title{Extract OSM contributions}
\usage{
ohsome_extract_contributions(
  boundary = NULL,
  geometryType = c("centroid", "bbox", "geometry"),
  latest = FALSE,
  time = lubridate::format_ISO8601(.ohsome_metadata$extractRegion$temporalExtent),
  properties = NULL,
  clipGeometry = TRUE,
  ...
)

ohsome_contributions_bbox(boundary = NULL, ...)

ohsome_contributions_centroid(boundary = NULL, ...)

ohsome_contributions_geometry(boundary = NULL, ...)
}
\arguments{
\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{geometryType}{Type of geometry to be extracted. Can be centroids
(\code{centroid}, bounding boxes (\code{bbox}) or \code{geometry}.
Default: \code{centroid}. Caveat: Node elements are omitted from results
in queries for bounding boxes.}

\item{latest}{logical; if true, request only the latest state of the
contributions provided to the OSM data.}

\item{time}{must consist of two ISO-8601 conform timestrings (as
comma-separated character or as character vector of length two) defining
a time interval (defaults to the temporal extent of the underlying OSHDB)}

\item{properties}{Can be "tags" to extract all tags and/or "metadata" to
provide metadata (\code{changesetId}, \code{lastEdit}, \code{osmType} and
\code{version}) and/or "contributionTypes" to provide contribution types
with the contributions. Multiple values can be provided as
comma-separated character or as character vector.
Default: NULL (provides \code{contributionChangesetId}, \code{osmId} and
\code{timestamp} with the contribution geometries)}

\item{clipGeometry}{logical; specifies whether the returned geometries of the
contributions should be clipped to the query’s spatial boundary}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for OSM contribution extraction
}
\details{
\code{ohsome_extract_contributionss} creates an \code{ohsome_query} object
for OSM contribution extraction. \code{ohsome_contributions_bbox},
\code{ohsome_contributions_centroid} and \code{ohsome_contributions_geometry}
are wrapper functions for specific contributions extraction endpoints.
Boundary objects are passed via \code{\link{set_boundary}} into
\code{\link{ohsome_boundary}}.
}
\examples{

# query to extract contributions to man-made features around "null island" with metadata:
ohsome_contributions_geometry("0,0,10", filter = "man_made=*", properties = "metadata")
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#contributions-extraction}
}
\concept{Extract contributions}
