% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_extract_elements.R
\name{ohsome_extract_elements}
\alias{ohsome_extract_elements}
\alias{ohsome_elements_bbox}
\alias{ohsome_elements_centroid}
\alias{ohsome_elements_geometry}
\title{Extract OSM elements}
\usage{
ohsome_extract_elements(
  boundary = NULL,
  geometryType = c("centroid", "bbox", "geometry"),
  properties = NULL,
  clipGeometry = TRUE,
  ...
)

ohsome_elements_bbox(boundary = NULL, ...)

ohsome_elements_centroid(boundary = NULL, ...)

ohsome_elements_geometry(boundary = NULL, ...)
}
\arguments{
\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{geometryType}{Type of geometry to be extracted. Can be centroids
(\code{centroid}, bounding boxes (\code{bbox}) or \code{geometry}.
Default: \code{centroid}. Caveat: Node elements are omitted from results
in queries for bounding boxes.}

\item{properties}{Can be "tags" to extract all tags with the elements and/or
"metadata" to provide metadata (\code{changesetId}, \code{lastEdit},
\code{osmType} and \code{version}) with the elements. Multiple values
can be provided as comma-separated character or as character vector.
Default: NULL (only provides \code{osmId} with the elements)}

\item{clipGeometry}{logical; specifies whether the returned geometries of the
features should be clipped to the query’s spatial boundary}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for OSM element extraction
}
\details{
\code{ohsome_extract_elements} creates an \code{ohsome_query} object for
OSM element extraction. \code{ohsome_elements_bbox},
\code{ohsome_elements_centroid} and \code{ohsome_elements_geometry} are
wrapper functions for specific elements extraction endpoints. Boundary
objects are passed via \code{\link{set_boundary}} into
\code{\link{ohsome_boundary}}.
}
\examples{

# query to extract geometries, metadata and tags of all elements with man_made tag around "null island":
ohsome_elements_geometry("0,0,10", filter = "man_made=*", properties = c("metadata", "tags"))
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#elements-extraction}
}
\concept{Extract elements}
