% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_extract_elementsFullHistory.R
\name{ohsome_extract_elementsFullHistory}
\alias{ohsome_extract_elementsFullHistory}
\alias{ohsome_elementsFullHistory_bbox}
\alias{ohsome_elementsFullHistory_centroid}
\alias{ohsome_elementsFullHistory_geometry}
\title{Extract OSM elements' full history}
\usage{
ohsome_extract_elementsFullHistory(
  boundary = NULL,
  geometryType = c("centroid", "bbox", "geometry"),
  time,
  properties = NULL,
  clipGeometry = TRUE,
  ...
)

ohsome_elementsFullHistory_bbox(boundary = NULL, ...)

ohsome_elementsFullHistory_centroid(boundary = NULL, ...)

ohsome_elementsFullHistory_geometry(boundary = NULL, ...)
}
\arguments{
\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{geometryType}{Type of geometry to be extracted. Can be centroids
(\code{centroid}, bounding boxes (\code{bbox}) or \code{geometry}.
Default: \code{centroid}. Caveat: Node elements are omitted from results
in queries for bounding boxes.}

\item{time}{must consist of two ISO-8601 conform timestrings (as
comma-separated character or as character vector of length two) defining
a time interval}

\item{properties}{Can be "tags" to extract all tags with the elements and/or
"metadata" to provide metadata (\code{changesetId}, \code{lastEdit},
\code{osmType} and \code{version}) with the elements. Multiple values
can be provided as comma-separated character or as character vector.
Default: NULL (only provides \code{osmId} with the elements)}

\item{clipGeometry}{logical Specifies whether the returned geometries of the
features should be clipped to the query’s spatial boundary}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for the extraction of OSM elements' full
history
}
\details{
\code{ohsome_extract_elementsFullHistory} creates an \code{ohsome_query}
object for OSM element full history extraction.
\code{ohsome_elementsFullHistory_bbox},
\code{ohsome_elementsFullHistory_centroid} and
\code{ohsome_elementsFullHistory_geometry} are wrapper functions for specific
elementsFullHistory extraction endpoints. Boundary objects are passed via
\code{\link{set_boundary}} into \code{\link{ohsome_boundary}}.
}
\examples{

# query to extract full history of building geometries around Heidelberg 
# main station:
ohsome_elementsFullHistory_geometry(
    boundary = "8.67542,49.40347,1000", 
    time = "2008,2020",
    filter = "building=* and type:way", 
    clipGeometry = FALSE
)
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/endpoints.html#elements-full-history-extraction}
}
\concept{Extract elements full History}
