% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohsome_metadata}
\alias{ohsome_metadata}
\title{ohsome API metadata}
\format{
An \code{ohsome_metadata} object. This is a named list with the
attributes \code{date} and \code{status_code} (of the GET request) and
the following list elements:
\describe{
\item{attribution}{\code{url} and \code{text} of OSM data copyrights
and attribution (character)}
\item{apiVersion}{Version of the ohsome API
(numeric_version)}
\item{timeout}{Limit of the processing time in seconds (numeric)}
\item{extractRegion}{
\code{spatialExtent}: {Spatial boundary of the OSM data in the
underlying OSHDB (sfc_POLYGON)}\cr
\code{temporalExtent}: {Timeframe of the OSM data in the
underlying OSHDB data (vector of POSIXct)}\cr
\code{replicationSequenceNumber}: {Precise state of the OSM data
contained in the underlying OSHDB, expressed as the id of the
last applied (hourly) diff file from
\url{planet.openstreetmap.org} (numeric)}
}}
}
\description{
Metadata of the ohsome API that is requested on loading the
package
}
