% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_parse.R
\name{ohsome_parse}
\alias{ohsome_parse}
\title{Parse content from an ohsome API response}
\usage{
ohsome_parse(response)
}
\arguments{
\item{response}{A response object}
}
\value{
A list (if the response is of type "application/json"), an sf
object (if the response is of type "application/geo+json") or a
data.frame (if the response is of type "text/csv")
}
\description{
Extract and parse the content from an ohsome API response
}
