% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_post.R
\name{ohsome_post}
\alias{ohsome_post}
\title{Send POST request to ohsome API}
\usage{
ohsome_post(
  query,
  parse = TRUE,
  validate = TRUE,
  additional_identifiers = NULL
)
}
\arguments{
\item{query}{an ohsome_query object constructed with \code{\link{ohsome_query}}
or any of its wrapper functions}

\item{parse}{logical parse the ohsome API response?}

\item{validate}{logical If true, issues warning for invalid endpoint or
invalid/missing query parameters}

\item{additional_identifiers}{optional user agent identifiers in addition to
"ohsome-r/version" (a vector coercible to character)}
}
\value{
\describe{
\item{an \code{sf} object}{if parse = TRUE and ohsome API response is GeoJSON}
\item{a \code{data.frame}}{if parse = TRUE and ohsome API response is not GeoJSON}
\item{an \code{ohsome_response} object}{if parse = FALSE}
}
}
\description{
Sends an ohsome_query object as a POST request to the ohsome API and silently
returns the unparsed response.
}
\examples{
q <- ohsome_elements_count(osmdata::getbb("Berlin"), filter = "amenity=cinema")

ohsome_post(q)
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
