% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_query.R
\name{ohsome_query}
\alias{ohsome_query}
\title{Create an \code{ohsome_query} object}
\usage{
ohsome_query(endpoint, ..., validate = FALSE)
}
\arguments{
\item{endpoint}{The path to the ohsome API endpoint. Either a single string
(e.g. \code{"elements/count"}) or a character vector in the right order
(e.g. \code{c("elements", "count")})}

\item{...}{Parameters of the request to the ohsome API endpoint}

\item{validate}{logical If TRUE, issues warning for invalid endpoint or
invalid/missing query parameters}
}
\value{
An \code{ohsome_query} object. The object consists of the following elements:
\describe{
\item{url}{The url of the endpoint.}
\item{encode}{The way the information is encoded then posted to the ohsome API. Set as "form".}
\item{body}{The parameters of the query such as \code{format},
\code{filter} or \code{bpolys}.}
}
}
\description{
Creates an \code{ohsome_query} object specifying the ohsome API endpoint and
the request parameters.
}
\examples{
ohsome_query("elements/geometry", bbox = "8.5992,49.3567,8.7499,49.4371", filter = "building=*")
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
