% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters.R
\name{set_parameters}
\alias{set_parameters}
\alias{set_time}
\alias{set_filter}
\alias{set_groupByKeys}
\alias{set_groupByKey}
\alias{set_groupByValues}
\alias{set_properties}
\title{Set parameters}
\usage{
set_parameters(query, ...)

set_time(query, time = NULL)

set_filter(query, filter = NULL, filter2 = NULL)

set_groupByKeys(query, groupByKeys = NULL)

set_groupByKey(query, groupByKey = NULL)

set_groupByValues(query, groupByValues = NULL)

set_properties(query, properties = NULL)
}
\arguments{
\item{query}{An \code{ohsome_query} object}

\item{...}{named parameters (e.g. \code{time = "2020-01-01"})}
}
\value{
An \code{ohsome_query} object
}
\description{
Set or modify parameters of an existing \code{ohsome_query} object
}
\details{
\code{set_parameters()} takes an \code{ohsome_query} object and an arbitrary
number of named parameters as an input. It sets or modifies these parameters
in the \code{ohsome_query} and returns the modified object. \code{set_time()},
\code{set_filter()}, \code{set_groupByKey()}, \code{set_groupByKey()},
\code{set_groupByValues()} and  \code{set_properties()} are wrapper functions
to set specific parameters. By default, an unmodified \code{ohsome_query}
object is returned.
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
\concept{Set parameters}
