% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_endpoint}
\alias{validate_endpoint}
\title{Validate endpoint}
\usage{
validate_endpoint(endpoint)
}
\arguments{
\item{endpoint}{The path to the ohsome API endpoint as a single string
(e.g. \code{"elements/count"})}
}
\value{
logical
}
\description{
Checks if the specified endpoint is in the list of known ohsome API endpoints
and issues a warning if not. Silently returns a logical that indicates the
validity of the endpoint.
}
\keyword{Internal}
