% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_json}
\alias{validate_json}
\title{Validate JSON response}
\usage{
validate_json(json)
}
\arguments{
\item{json}{A JSON text string}
}
\value{
logical
}
\description{
Validates ohsome API response content of type (Geo)JSON. Specifically checks
for HTTP status messages in the content and uses any message to issue a
warning. Returns a logical that indicates the validity of the JSON.
}
\keyword{Internal}
