% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate parameters}
\usage{
validate_parameters(endpoint, body)
}
\arguments{
\item{endpoint}{The path to the ohsome API endpoint as a single string
(e.g. \code{"elements/count"})}

\item{body}{A list of named parameters to the ohsome API request}
}
\value{
logical
}
\description{
Checks if the specified parameters in a request body are valid and issues a
warning if not. Returns a logical that indicates the validity of the
parameters.
}
\keyword{Internal}
