% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_content}
\alias{convert_content}
\title{Convert ohsome metadata content}
\usage{
convert_content(content)
}
\arguments{
\item{content}{The parsed content of a response from the metadata endpoint of
ohsome API}
}
\value{
a list
}
\description{
Convert ohsome metadata content
}
\keyword{internal}
