% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_endpoint}
\alias{extract_endpoint}
\title{Extract endpoint}
\usage{
extract_endpoint(query)
}
\arguments{
\item{query}{an ohsome_query object}
}
\value{
character
}
\description{
Extract the API endpoint path from the URL in an ohsome_query object
}
\keyword{Internal}
