% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_aggregate_elements.R
\name{ohsome_aggregate_elements}
\alias{ohsome_aggregate_elements}
\alias{ohsome_elements_count}
\alias{ohsome_elements_length}
\alias{ohsome_elements_perimeter}
\alias{ohsome_elements_area}
\title{Aggregate OSM elements}
\usage{
ohsome_aggregate_elements(
  boundary = NULL,
  aggregation = c("count", "length", "perimeter", "area"),
  return_value = c("absolute", "density"),
  grouping = NULL,
  ...
)

ohsome_elements_count(boundary = NULL, ...)

ohsome_elements_length(boundary = NULL, ...)

ohsome_elements_perimeter(boundary = NULL, ...)

ohsome_elements_area(boundary = NULL, ...)
}
\arguments{
\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{aggregation}{Aggregation operation}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for OSM element aggregation
}
\details{
\code{ohsome_aggregate_elements} creates an \code{ohsome_query} object for
OSM element aggregation. \code{ohsome_elements_count},
\code{ohsome_elements_length}, \code{ohsome_elements_perimeter} and
\code{ohsome_elements_area} are wrapper functions for specific aggregation
endpoints. Boundary objects are passed via \code{\link{set_boundary}} into
\code{\link{ohsome_boundary}}.
}
\examples{

ohsome_aggregate_elements(mapview::franconia, "count", filter = "craft=brewery")
ohsome_elements_count(mapview::franconia, filter = "craft=brewery")
ohsome_elements_length(mapview::franconia, filter = "highway=*")

}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#elements-aggregation}
}
\concept{Aggregate elements}
