% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_aggregate_elements.R
\name{ohsome_aggregate_elements}
\alias{ohsome_aggregate_elements}
\alias{ohsome_elements_count}
\alias{ohsome_elements_length}
\alias{ohsome_elements_perimeter}
\alias{ohsome_elements_area}
\title{Aggregate OSM elements}
\usage{
ohsome_aggregate_elements(
  boundary = NULL,
  aggregation = c("count", "length", "perimeter", "area"),
  return_value = c("absolute", "density", "ratio"),
  grouping = NULL,
  time = lubridate::format_ISO8601(ohsome_temporalExtent[2]),
  ...
)

ohsome_elements_count(boundary = NULL, ...)

ohsome_elements_length(boundary = NULL, ...)

ohsome_elements_perimeter(boundary = NULL, ...)

ohsome_elements_area(boundary = NULL, ...)
}
\arguments{
\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{aggregation}{character; aggregation operation}

\item{return_value}{character; the value to be returned by the ohsome API:
\describe{
\item{absolute}{Returns the absolute number, length, perimeter or area
of elements (default).}
\item{density}{Returns the number, length, perimeter or area
of elements per square kilometer.}
\item{ratio}{Returns an absolute \code{value} for elements satisfying
the \code{filter} argument, an absolute \code{value2} for elements
satisfying the \code{filter2} argument, and the \code{ratio} of
\code{value2} to \code{value}}
}}

\item{grouping}{character; group type(s) for grouped aggregations. The
following group types are available:
\describe{
\item{boundary}{Groups the result by the given boundaries that are defined through any of the boundary query parameters}
\item{key}{Groups the result by the given keys that are defined through the groupByKeys query parameter.}
\item{tag}{Groups the result by the given tags that are defined through the groupByKey and groupByValues query parameters.}
\item{type}{Groups the result by the given OSM, or simple feature types that are defined through the types parameter.}
\item{c("boundary", "tag")}{Groups the result by the given boundary and the tags.}
}
Not all of these group types are accepted by all of the aggregation
endpoints. Check
\href{https://docs.ohsome.org/ohsome-api/v1/group-by.html}{Grouping}
for available group types.}

\item{time}{character; time parameter of the query (see
\href{https://docs.ohsome.org/ohsome-api/v1/time.html}{Supported time formats};
defaults to most recent available timestamp in the underlying OSHDB)}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for OSM element aggregation
}
\details{
\code{ohsome_aggregate_elements} creates an \code{ohsome_query} object for
OSM element aggregation. \code{ohsome_elements_count},
\code{ohsome_elements_length}, \code{ohsome_elements_perimeter} and
\code{ohsome_elements_area} are wrapper functions for specific aggregation
endpoints. Boundary objects are passed via \code{\link{set_boundary}} into
\code{\link{ohsome_boundary}}.
}
\examples{

ohsome_aggregate_elements(mapview::franconia, "count", filter = "craft=brewery")
ohsome_elements_count(mapview::franconia, filter = "craft=brewery")
ohsome_elements_length(mapview::franconia, filter = "highway=*")

}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#elements-aggregation}
}
\concept{Aggregate elements}
