% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohsome_api_url}
\alias{ohsome_api_url}
\title{ohsome API URL}
\format{
A list:
\itemize{
\item \code{base}: character; base URL
\item \code{version}: character; path to current major API version
}
}
\description{
The base URL of the ohsome API with path to current major
version.
}
