% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_boundary.R
\name{ohsome_boundary}
\alias{ohsome_boundary}
\alias{ohsome_boundary.ohsome_boundary}
\alias{ohsome_boundary.character}
\alias{ohsome_boundary.sf}
\alias{ohsome_boundary.sfc}
\alias{ohsome_boundary.sfg}
\alias{ohsome_boundary.bbox}
\alias{ohsome_boundary.matrix}
\alias{ohsome_boundary.list}
\title{Create \code{ohsome_boundary} object}
\usage{
ohsome_boundary(x, ...)

\method{ohsome_boundary}{ohsome_boundary}(x, ...)

\method{ohsome_boundary}{character}(x, ...)

\method{ohsome_boundary}{sf}(x, digits = 6, ...)

\method{ohsome_boundary}{sfc}(x, ...)

\method{ohsome_boundary}{sfg}(x, ...)

\method{ohsome_boundary}{bbox}(x, ...)

\method{ohsome_boundary}{matrix}(x, ...)

\method{ohsome_boundary}{list}(x, ...)
}
\arguments{
\item{x}{Bounding geometries that are passed to
\code{\link{ohsome_boundary}}. Bounding geometries should be in WGS84 in the format lon-lat. For \code{sf} objects the geometries are transformed to WGS84 if the CRS of the object is known. The following classes are supported:
\describe{
\item{sf}{with (MULTI)POLYGON geometries}
\item{sfc}{with (MULTI)POLYGON geometries}
\item{sfg}{with (MULTI)POLYGON geometries and WGS 84 coordinates}
\item{bbox}{created with \code{\link[sf]{st_bbox}}}
\item{matrix}{created with
\code{\link[sp]{bbox}} or
\code{\link[osmdata]{getbb}}
}
\item{character}{a textual definition of bounding polygons, boxes or
circles as allowed by the ohsome API (see
\href{https://docs.ohsome.org/ohsome-api/stable/boundaries.html}{documentation})
}
\item{list}{a list of \code{bbox} objects, \code{matrix} or
\code{character}. Bounding geometry types of all list elements must
be the same. Does not work with GeoJSON FeatureCollections.
}
}}

\item{...}{Additional arguments other than \code{digits} are ignored.}

\item{digits}{Number of decimal places of coordinates in the resulting
GeoJSON when converting \code{sf} to GeoJSON (defaults to 6).}
}
\value{
An \code{ohsome_boundary} object which contains the following elements: \code{boundary} which
contains the boundary in geojson format and the \code{type} of the boundary (bpolys, bcircles, or bboxes).
}
\description{
Creates an \code{ohsome_boundary} object from input geometries of various
classes. The \code{ohsome_boundary} object is used to set the \code{bpolys}, \code{bboxes} or
\code{bcircles} parameters of an \code{ohsome_query} object.
}
\examples{
# define boundary by a circle (bcircle)
ohsome_boundary("8.6528,49.3683,1000") 
# define boundary by two circles (named Circle 1 and Circle 2)
ohsome_boundary("Circle 1:8.6528,49.3683,1000|Circle 2:8.7294,49.4376,1000")
# use the shape of franconia from the mapview package as bpolys
ohsome_boundary(mapview::franconia, digits = 4)
# get administrative boundary for Berlin from OSM and use as bpolys
ohsome_boundary(osmdata::getbb("Berlin"))
# use the bounding box from franconia
ohsome_boundary(sf::st_bbox(mapview::franconia))
# use a list with two bounding boxes
ohsome_boundary(list(osmdata::getbb("Berlin"), sf::st_bbox(mapview::franconia)))
}
