% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_extract_elements.R
\name{ohsome_extract_elements}
\alias{ohsome_extract_elements}
\alias{ohsome_elements_bbox}
\alias{ohsome_elements_centroid}
\alias{ohsome_elements_geometry}
\title{Extract OSM elements}
\usage{
ohsome_extract_elements(
  boundary = NULL,
  geometryType = c("centroid", "bbox", "geometry"),
  time = lubridate::format_ISO8601(ohsome_temporalExtent[2]),
  properties = NULL,
  clipGeometry = TRUE,
  ...
)

ohsome_elements_bbox(boundary = NULL, ...)

ohsome_elements_centroid(boundary = NULL, ...)

ohsome_elements_geometry(boundary = NULL, ...)
}
\arguments{
\item{boundary}{Bounding geometries specified in WGS84 in the format lon-lat.
The geometries of \code{sf} are transformed to WGS84 if the CRS of the object is
known. The following classes are supported:
\itemize{
\item \code{sf} with (MULTI)POLYGON geometries
\item \code{sfc} with (MULTI)POLYGON geometries
\item \code{sfg} with (MULTI)POLYGON geometries and WGS 84 coordinates
\item \code{bbox} created with \code{\link[sf:st_bbox]{sf::st_bbox()}} or \code{\link[tmaptools:bb]{tmaptools::bb()}}
\item \code{matrix} created with \code{\link[sp:bbox]{sp::bbox()}} or \code{\link[osmdata:getbb]{osmdata::getbb()}}
\item \code{character} providing textual definitions of bounding polygons, boxes or
circles as allowed by the ohsome API (see
\href{https://docs.ohsome.org/ohsome-api/stable/boundaries.html}{ohsome API - Boundaries})
\item \code{list} of \code{bbox}, \code{matrix} or \code{character}. Bounding geometry types of all
list elements must be the same. Does not work with GeoJSON
FeatureCollections.
}}

\item{geometryType}{character; type of geometry to be extracted:
\itemize{
\item \code{"centroid"},
\item \code{"bboxes"} (bounding boxes), or
\item \code{"geometry"}
}

Caveat: Node elements are omitted from results in queries for bounding
boxes.}

\item{time}{character; \code{time} parameter of the query (see
\href{https://docs.ohsome.org/ohsome-api/v1/time.html}{Supported time formats}).
This defaults to the most recent available timestamp in the underlying
OSHDB.}

\item{properties}{character; properties to be extracted with the features:
\itemize{
\item \code{"tags"}, and/or
\item \code{"metadata"} (i.e. \verb{@changesetId}, \verb{@lastEdit}, \verb{@osmType}, and
\verb{@version})
}

Multiple values can be provided as comma-separated character or as
character vector. This defaults to \code{NULL} (provides \verb{@osmId}).}

\item{clipGeometry}{logical; specifies whether the returned geometries should
be clipped to the query’s spatial boundary}

\item{...}{Parameters of the request to the ohsome API endpoint.}
}
\value{
An \code{ohsome_query} object. The object can be sent to the ohsome API
with \code{\link[=ohsome_post]{ohsome_post()}}. It consists of the following elements:
\itemize{
\item \code{url}: The URL of the endpoint.
\item \code{encode}: The way the information is encoded and then posted to the
ohsome API. Set as \code{"form"}.
\item \code{body}: The parameters of the query such as \code{format}, \code{filter} or
\code{bpolys}.
}
}
\description{
Create an \code{ohsome_query} object for OSM element extraction
}
\details{
\code{ohsome_extract_elements()} creates an \code{ohsome_query} object for OSM element
extraction. \code{ohsome_elements_bbox()}, \code{ohsome_elements_centroid()} and
\code{ohsome_elements_geometry()} are wrapper functions for specific elements
extraction endpoints. Boundary objects are passed via \code{\link[=set_boundary]{set_boundary()}} into
\code{\link[=ohsome_boundary]{ohsome_boundary()}}.
}
\examples{
# Extract geometries, metadata and tags man-made objects around "Null Island":
ohsome_elements_geometry(
    "0,0,10", 
    filter = "man_made=*", 
    time = "2022-01-01",
    properties = c("metadata", "tags")
)

}
\seealso{
\href{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#elements-extraction}{ohsome API Endpoints -- Elements Extraction}
}
\concept{Extract elements}
