% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_get_metadata.R
\name{ohsome_get_metadata}
\alias{ohsome_get_metadata}
\title{GET metadata from ohsome API}
\usage{
ohsome_get_metadata()
}
\value{
a list of metadata (see \code{\link{ohsome_metadata}})
}
\description{
\code{ohsome_get_metadata} returns the parsed metadata of ohsome API
}
\details{
\code{ohsome_get_metadata} sends a GET request to the metadata endpoint of
ohsome API and parses the response. The parsed metadata is returned as well
assigned to \code{\link{ohsome_metadata}} as a side effect.
}
\examples{
\dontrun{
ohsome_get_metadata()
}
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/endpoints.html#metadata}
}
