% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_get_metadata.R
\name{ohsome_get_metadata}
\alias{ohsome_get_metadata}
\title{GET metadata from ohsome API}
\usage{
ohsome_get_metadata(quiet = FALSE)
}
\arguments{
\item{quiet}{logical; suppress message on data attribution, API version and
temporal extent.}
}
\value{
An \code{ohsome_metadata} object. This is a named list with the
attributes \code{date} and \code{status_code} (of the GET request) and
the following list elements:
\describe{
\item{attribution}{\code{url} and \code{text} of OSM data copyrights
and attribution (character)}
\item{apiVersion}{Version of the ohsome API
(numeric_version)}
\item{timeout}{Limit of the processing time in seconds (numeric)}
\item{extractRegion}{
\code{spatialExtent}: {Spatial boundary of the OSM data in the
underlying OSHDB (sfc_POLYGON)}\cr
\code{temporalExtent}: {Timeframe of the OSM data in the
underlying OSHDB data (vector of POSIXct)}\cr
\code{replicationSequenceNumber}: {Precise state of the OSM data
contained in the underlying OSHDB, expressed as the id of the
last applied (hourly) diff file from
\link{planet.openstreetmap.org} (numeric)}
}}
}
\description{
Return parsed metadata from ohsome API
}
\details{
\code{ohsome_get_metadata} sends a GET request to the metadata endpoint of
ohsome API and parses the response. The parsed metadata is silently returned.
}
\examples{
\dontrun{
ohsome_get_metadata()
}
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/endpoints.html#metadata}
}
