% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_get_metadata.R
\name{ohsome_get_metadata}
\alias{ohsome_get_metadata}
\title{GET metadata from ohsome API}
\usage{
ohsome_get_metadata()
}
\value{
a list of metadata (see \code{\link{ohsome_metadata}})
}
\description{
Sends a GET request to the metadata endpoint of ohsome API, parses the
response and assigns it to ohsome_metadata
}
\examples{
\dontrun{
ohsome_get_metadata()
}
}
