% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohsome_metadata}
\alias{ohsome_metadata}
\title{ohsome API metadata}
\format{
An \code{ohsome_metadata} object. This is a named list with the
attributes \code{date}, \code{status_code} (of the GET request) and the following list
elements:
\itemize{
\item \code{attribution}: characer; \code{url} and \code{text} of OSM data copyrights and
attribution
\item \code{apiVersion}: numeric_version; Version of the ohsome API
\item \code{timeout}: numeric; limit of the processing time in seconds
\item \code{extractRegion}:
\itemize{
\item \code{spatialExtent}: sfc_POLYGON; spatial boundary of the OSM data in the
underlying OSHDB
\item \code{temporalExtent}: vector of POSIXct; timeframe of the OSM data in the
underlying OSHDB data
\item \code{replicationSequenceNumber}: numeric; precise state of the OSM data
contained in the underlying OSHDB, expressed as the id of the last
applied (hourly) diff file from \href{planet.openstreetmap.org}{Planet OSM}
}
}
}
\description{
Metadata of the ohsome API that is requested on loading the
package
}
