% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_post.R
\name{ohsome_post}
\alias{ohsome_post}
\title{Send POST request to ohsome API}
\usage{
ohsome_post(ohsome_query, parse = TRUE, additional_identifiers = NULL)
}
\arguments{
\item{ohsome_query}{an ohsome_query object constructed with ohsome_query()
or any of its wrapper functions}

\item{parse}{logical parse the ohsome API response?}

\item{additional_identifiers}{optional user agent identifiers in addition to
"ohsome-r/version" (a vector coercible to character)}
}
\value{
\describe{
\item{an sf object}{if parse = TRUE and ohsome API response is GeoJSON}
\item{a data.frame}{if parse = TRUE and ohsome API response is not GeoJSON}
\item{an ohsome_response object}{if parse = FALSE}
}
}
\description{
Sends an ohsome_query object as a POST request to the ohsome API and silently
returns the unparsed response.
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
