% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_post.R
\name{ohsome_post}
\alias{ohsome_post}
\title{Send POST request to ohsome API}
\usage{
ohsome_post(ohsome_query)
}
\arguments{
\item{ohsome_query}{}
}
\value{

}
\description{
Send POST request to ohsome API
}
