% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_query.R
\name{ohsome_query}
\alias{ohsome_query}
\title{Create an ohsome_query object}
\usage{
ohsome_query(endpoint, ...)
}
\arguments{
\item{endpoint}{The path to the ohsome API endpoint. Either a single string
(e.g. \code{"elements/count"}) or a character vector in the right order
(e.g. \code{c("elements", "count")})}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Creates an ohsome_query object specifying the ohsome API endpoint and the
request parameters.
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
