% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_query.R
\name{ohsome_query}
\alias{ohsome_query}
\title{Create an \code{ohsome_query} object}
\usage{
ohsome_query(endpoint, boundary = NULL, grouping = NULL, ..., validate = FALSE)
}
\arguments{
\item{endpoint}{The path to the ohsome API endpoint. Either a single string
(e.g. \code{"elements/count"}) or a character vector in the right order
(e.g. \code{c("elements", "count")})}

\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{grouping}{character Group type(s) for grouped aggregations (only
available for queries to aggregation endpoints). In general, the following
group types are available:
\describe{
\item{"boundary"}{Groups the result by the given boundaries that are defined through any of the boundary query parameters}
\item{"key"}{Groups the result by the given keys that are defined through the groupByKeys query parameter.}
\item{"tag"}{Groups the result by the given tags that are defined through the groupByKey and groupByValues query parameters.}
\item{"type"}{Groups the result by the given OSM, or simple feature types that are defined through the types parameter.}
\itme{c("boundary", "tag")}{Groups the result by the given boundary and the tags. Provide }
}
Not all of these group types are accepted by all of the aggregation
endpoints.Please consult the ohsome API documentation to check for available
group types.}

\item{...}{Parameters of the request to the ohsome API endpoint}

\item{validate}{logical If TRUE, issues warning for invalid endpoint or
invalid/missing query parameters}
}
\value{
An \code{ohsome_query} object. The object consists of the following elements:
\describe{
\item{url}{The url of the endpoint.}
\item{encode}{The way the information is encoded then posted to the ohsome API. Set as "form".}
\item{body}{The parameters of the query such as \code{format},
\code{filter} or \code{bpolys}.}
}
}
\description{
Creates an \code{ohsome_query} object specifying the ohsome API endpoint and
the request parameters.
}
\examples{
# setting bbox parameter manually
ohsome_query("elements/geometry", bboxes = "8.6,49.36,8.75,49.44", filter = "building=*")

# using a boundary object:
ohsome_query("elements/geometry", boundary = "8.6,49.36,8.75,49.44", filter = "building=*")
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
