% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsome_users_count.R
\name{ohsome_users_count}
\alias{ohsome_users_count}
\title{Count OSM users}
\usage{
ohsome_users_count(
  boundary = NULL,
  return_value = c("absolute", "density"),
  grouping = NULL,
  ...
)
}
\arguments{
\item{boundary}{Boundary object that can be interpreted by
\code{\link{ohsome_boundary}}}

\item{return_value}{character; the value to be returned by the ohsome API:
\describe{
\item{absolute}{Returns the absolute number of users (default).}
\item{density}{Returns the number of users per square kilometer.}
}}

\item{grouping}{character; group type(s) for grouped user aggregation. The
following group types are available:
\describe{
\item{boundary}{Groups the result by the given boundaries that are defined through any of the boundary query parameters}
\item{key}{Groups the result by the given keys that are defined through the groupByKeys query parameter.}
\item{tag}{Groups the result by the given tags that are defined through the groupByKey and groupByValues query parameters.}
\item{type}{Groups the result by the given OSM, or simple feature types that are defined through the types parameter.}
}
Not all of these group types are accepted by all of the user aggregation
endpoints. Please consult the
\href{https://docs.ohsome.org/ohsome-api/v1/group-by.html}{ohsome API documentation}
to check for available group types.}

\item{...}{Parameters of the request to the ohsome API endpoint}
}
\value{
An \code{ohsome_query} object
}
\description{
Create an \code{ohsome_query} object for OSM users count
}
\details{
\code{ohsome_users_count} creates an \code{ohsome_query} object for
OSM users aggregation. Boundary objects are passed via \code{\link{set_boundary}} into
\code{\link{ohsome_boundary}}.
}
\examples{

ohsome_users_count("0,0,10", filter = "man_made=*", time = "2010/2020/P1Y")

}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/stable/endpoints.html#users-aggregation}
}
