% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_content}
\alias{parse_content}
\title{Parse content from a response}
\usage{
parse_content(resp)
}
\arguments{
\item{resp}{A response object}
}
\value{
A list (if the response is of type "application/json") or an sf
object (if the response is of type "application/geo+json")
}
\description{
Extract and parse the content from an ohsome API response and handle API
errors.
}
\keyword{Internal}
