% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_content}
\alias{parse_content}
\title{Parse content from a response}
\usage{
parse_content(resp)
}
\arguments{
\item{resp}{A response object}
}
\value{
A list (parsed and converted content of ohsome metadata)
}
\description{
A wrapper around \code{\link{httr::content}} to extract and parse the content
from an API response and handle API errors.
}
\keyword{Internal}
