% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_boundary.R
\name{set_boundary}
\alias{set_boundary}
\alias{set_bpolys}
\alias{set_bboxes}
\alias{set_bcircles}
\title{Set boundary}
\usage{
set_boundary(
  query,
  boundary,
  boundary_type = c("bpolys", "bboxes", "bcircles")
)

set_bpolys(query, boundary)

set_bboxes(query, boundary)

set_bcircles(query, boundary)
}
\arguments{
\item{query}{An \code{ohsome_query} object}

\item{boundary_type}{}
}
\value{
An \code{ohsome_query} object
}
\description{
Modify the spatial filter of an existing \code{ohsome_query} object
}
\details{
\code{set_boundary()} adds or modifies aspatial filter of an
\code{ohsome_query} object. The spatial filter of a query to the ohsome API
can be defined as dne or more polygons, bounding boxes or bounding circles.
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
