% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_endpoint.R
\name{set_endpoint}
\alias{set_endpoint}
\title{Set endpoint}
\usage{
set_endpoint(query, endpoint, append = FALSE, reset_format = TRUE)
}
\arguments{
\item{query}{An \code{ohsome_query} object}

\item{endpoint}{The path to the ohsome API endpoint. Either a single string
(e.g. \code{"elements/count"}) or a character vector in the right order
(e.g. \code{c("elements", "count")})}

\item{append}{If true, the provided endpoint string is appended to  the
existing endpoint definition instead of replacing it. This is
particularly useful if you wish to add a groupBy to an existing
aggregation query.}

\item{reset_format}{If true, the format parameter of the query will be
automatically determined depending on the new endpoint.}
}
\value{
An \code{ohsome_query} object
}
\description{
Modify the endpoint of an existing \code{ohsome_query} object
}
\details{
\code{set_endpoint()} takes an \code{ohsome_query} object and modifies the
ohsome API endpoint.
}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
