% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters.R
\name{set_parameters}
\alias{set_parameters}
\alias{set_time}
\alias{set_filter}
\alias{set_groupByKeys}
\alias{set_groupByKey}
\alias{set_groupByValues}
\alias{set_properties}
\title{Set parameters}
\usage{
set_parameters(query, ...)

set_time(query, time = NULL)

set_filter(query, filter = NULL, filter2 = NULL)

set_groupByKeys(query, groupByKeys = NULL)

set_groupByKey(query, groupByKey = NULL)

set_groupByValues(query, groupByValues = NULL)

set_properties(query, properties = NULL)
}
\arguments{
\item{query}{An \code{ohsome_query} object}

\item{...}{named parameters (e.g. \code{time = "2020-01-01"})}

\item{time}{time parameter of the query}

\item{filter}{filter parameter of the query}

\item{filter2}{filter2 parameter of a ratio query}

\item{groupByKeys}{groupByKeys parameter of a groupBy/key query}

\item{groupByKey}{groupByKey parameter of a groupBy/tag query}

\item{groupByValues}{groupByValues parameter of a groupBy/tag query}

\item{properties}{properties parameter of an extraction query. Can be
"tags" to extract all tags with the elements and/or
"metadata" to provide metadata with the elements. Multiple values
can be provided as comma-separated character or as character vector.
Default: NULL (removes \code{properties} parameter from query body)}
}
\value{
An \code{ohsome_query} object
}
\description{
Set or modify parameters of an existing \code{ohsome_query} object
}
\details{
\code{set_parameters()} takes an \code{ohsome_query} object and an arbitrary
number of named parameters as an input. It sets or modifies these parameters
in the \code{ohsome_query} and returns the modified object. \code{set_time()},
\code{set_filter()}, \code{set_groupByKey()}, \code{set_groupByKey()},
\code{set_groupByValues()} and  \code{set_properties()} are wrapper functions
to set specific parameters. By default, an unmodified \code{ohsome_query}
object is returned (except for \code{set_properties()} which removes the
\code{properties} parameter from the query by default).
}
\examples{
q <- ohsome_query("elements/count/ratio/groupBy/boundary")

q |>
    set_boundary("HD:8.5992,49.3567,8.7499,49.4371|HN:9.1638,49.113,9.2672,49.1766") |>
    set_time("2021") |>
    set_filter("building=*", filter2 = "building=* and building:levels=3") |>
    set_parameters(format = "csv")

}
\seealso{
\url{https://docs.ohsome.org/ohsome-api/v1/}
}
\concept{Set parameters}
