% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_query}
\alias{validate_query}
\title{Validate ohsome_query}
\usage{
validate_query(query)
}
\arguments{
\item{query}{an ohsome_query object constructed with ohsome_query()
or any of its wrapper functions}
}
\value{
logical
}
\description{
Validates an ohsome_query object by checking against ohsome_endpoints. Returns
a logical that indicates the validity of the query.
}
\keyword{Internal}
